Sort tracks in album by discNr and trackNr
The MPD server sends tracks in order of date by default when running a
find album "Album name" command, but not all MPD servers send it in that
order. Notably, OwnTone[1] sends them in alphabetical order of the track
title. In the protocol documentation the default sort order is
unspecified[2].

To sort tracks in a standard way across different MPD servers,
tracks will now be sorted in the client after receiving them
by disc number and then track number. This seems to be the approach
in the MPD clients Cantata and ncmpcpp as well.

To sort by disc number, the discnr property has been added to MpdTrack
and MpdFileEntry. Perhaps this can be used in the UI later.

Since MPD version 0.21, the MPD server supports a "sort" parameter for the
"find" command, but this is neither supported in OwnTone or Mopidy-MPD[3],
henche the choice for a client-side sort.

[1]: https://owntone.github.io/owntone-server/
[2]: https://mpd.readthedocs.io/en/latest/protocol.html#command-find
[3]: https://mopidy.com/ext/mpd/